Task 14: Balance factor

+++ 14-binary_tree_balance.c
@@ -6,7 +6,7 @@
  *
  * Return:The height of the tree or 0 if tree is NULL
  */
-int binary_tree_height(const binary_tree_t *tree)
+size_t binary_tree_height(const binary_tree_t *tree)
 {
 	if (tree == NULL)
 		return (0);
